Separate ice models and collect in SeaIce submodule
The three 0D ice models are now in separate files
Functions reused by all three are in shared.jl
Created a submodule for sea ice (and finally learned how modules work in Julia)
