+++ src/clima_playground.jl
@@ -7,7 +7,7 @@ include("analysis.jl")
 include("diagnostics.jl")
 include("components/atmosphere.jl")
 include("components/ocean.jl")
-include("components/ice.jl")
+include("components/SeaIce.jl")
 include("coupled_simulation.jl")
 include("monin_obukhov.jl")
 include("postprocessing.jl")
