+++ src/components/ocean.jl
@@ -4,6 +4,7 @@ import ClimaTimeSteppers as CTS
 import ClimaCoupler: Checkpointer, Interfacer
 import ClimaDiagnostics as CD
 import ClimaCore.MatrixFields: @name, ⋅, FieldMatrixWithSolver, FieldMatrix
+import LinearAlgebra
 import Statistics: mean
 
 export HeatEquationOcean, heat_oce_rhs!, ocean_init
